requirements.txt
gunicorn
numpy
flask
#tensorflow
opencv-python 
matplotlib
easyocr
pymongo
flask_cors
pillow